+++ packages/view-types/feature-list/src/FeatureListSubscriber.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import plur from 'plur';
 import { capitalize, commaNumber } from '@vitessce/utils';
 import {
@@ -9,6 +9,7 @@ import {
 } from '@vitessce/vit-s';
 import { ViewType, COMPONENT_COORDINATION_TYPES } from '@vitessce/constants-internal';
 import FeatureList from './FeatureList.js';
+import FeatureListOptions from './FeatureListOptions.js';
 
 
 /**
@@ -27,9 +28,10 @@ import FeatureList from './FeatureList.js';
  * @param {boolean} props.enableMultiSelect If true, allow
  * shift-clicking to select multiple genes.
  * @param {boolean} props.showTable If true, shows a table with the feature name and id.
- * @param {string} props.sort The sort order of the genes. If sort is defined and
+ * @param {'alphabetical'|'original'} props.sort The sort order of the genes. If sort is defined and
  * it is not equal to `alphabetical`, the genes will be displayed in the feature list in
  * the original order.
+ * @param {'featureIndex'|'featureLabels'|null} props.sortKey The information to use for sorting.
  */
 export function FeatureListSubscriber(props) {
   const {
@@ -41,6 +43,7 @@ export function FeatureListSubscriber(props) {
     enableMultiSelect = false,
     showTable = false,
     sort = 'alphabetical',
+    sortKey = null,
   } = props;
 
   const loaders = useLoaders();
@@ -81,11 +84,18 @@ export function FeatureListSubscriber(props) {
   ]);
   const geneList = featureIndex || [];
   const numGenes = geneList.length;
+  const hasFeatureLabels = Boolean(featureLabelsMap);
 
   function setGeneSelectionAndColorEncoding(newSelection) {
     setGeneSelection(newSelection);
     setCellColorEncoding('geneSelection');
   }
+  const [showFeatureTable, setShowFeatureTable] = useState(showTable);
+  const [featureListSort, setFeatureListSort] = useState(sort);
+  const [featureListSortKey, setFeatureListSortKey] = useState(null);
+  const initialSortKey = sortKey || (hasFeatureLabels ? 'featureLabels' : 'featureIndex');
+
+  const primaryColumnName = `${capitalize(featureType)} ID`;
 
   return (
     <TitleInfo
@@ -99,12 +109,25 @@ export function FeatureListSubscriber(props) {
       removeGridComponent={removeGridComponent}
       isReady={isReady}
       urls={urls}
+      options={(
+        <FeatureListOptions
+          featureListSort={featureListSort}
+          setFeatureListSort={setFeatureListSort}
+          featureListSortKey={featureListSortKey || initialSortKey}
+          setFeatureListSortKey={setFeatureListSortKey}
+          showFeatureTable={showFeatureTable}
+          setShowFeatureTable={setShowFeatureTable}
+          hasFeatureLabels={Boolean(featureLabelsMap)}
+          primaryColumnName={primaryColumnName}
+        />
+      )}
     >
       <FeatureList
         hasColorEncoding={cellColorEncoding === 'geneSelection'}
-        showTable={showTable}
+        showFeatureTable={showFeatureTable}
         geneList={geneList}
-        sort={sort}
+        featureListSort={featureListSort}
+        featureListSortKey={featureListSortKey || initialSortKey}
         featureLabelsMap={featureLabelsMap}
         featureType={featureType}
         geneSelection={geneSelection}
@@ -113,6 +136,8 @@ export function FeatureListSubscriber(props) {
         setGeneFilter={setGeneFilter}
         setGeneHighlight={setGeneHighlight}
         enableMultiSelect={enableMultiSelect}
+        hasFeatureLabels={Boolean(featureLabelsMap)}
+        primaryColumnName={primaryColumnName}
       />
     </TitleInfo>
   );